Language.Haskell.Meta.Extensions

Description

Documentation

toExtension :: KnownExtension -> Maybe Extension

Returns Nothing when the extension is not supported by template-haskell.

fromExtension :: Extension -> Maybe KnownExtension

Returns Nothing when the extension is not supported by haskell-src-exts.